Switches to React-Bootstrap, add supplier working
