+++ store-frontend/package.json
@@ -13,6 +13,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
+    "react-select-country-list": "^2.2.3",
     "react-validation": "^3.0.7",
     "validator": "^13.7.0",
     "web-vitals": "^1.0.1"
